django/seminar5/server/requirements.txt
asgiref==3.4.1
Django==3.2.6
pytz==2021.1
sqlparse==0.4.1
django-debug-toolbar==2.2
djangorestframework==3.11.1
djangorestframework-jwt
django-rest-authtoken
django-filter
factory-boy